rename spotify_track, use get_display_name more, fix a few errors
