+++ djyosof/cogs/spotify.py
@@ -29,7 +29,7 @@ class SpotifyCog(commands.Cog):
 
         tracklist_markdown = ""
         for idx, track in enumerate(tracks):
-            tracklist_markdown += f"**{idx+1}**. {track.name} - {track.artist}\n"
+            tracklist_markdown += f"**{idx+1}**. {track.get_display_name()}\n"
 
         embed.add_field(name="Search Results", value=tracklist_markdown)
