+++ djyosof/players/audio_player.py
@@ -64,7 +64,7 @@ class AudioPlayer:
                 after=lambda _: self.bot.loop.call_soon_threadsafe(self.next.set),
             )
 
-            print(f"Playing {track.name} by {track.artist}")
+            print(f"Playing {track.get_display_name()}")
             await interaction.followup.send(
                 content="Playing music!", embed=track.get_embed()
             )
